isolate libs in adapters (ftp / sftp / webdav)
